third_party/blink/web_tests/platform/generic/http/tests/websocket/unicode-expected.txt

WebSocket unicode message test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Connected.
Send UNICODE_HELLO.
PASS data is UNICODE_GOODBYE
Closed.
PASS successfullyParsed is true

TEST COMPLETE
